feat: Add components for displaying reports and notices
This commit adds two new components, Reports and Notices, for displaying data fetched from a GraphQL API. The Reports component fetches and displays a list of reports, while the Notices component fetches and displays a list of notices. These components will be used to show relevant information to the user.